added mapping, write and mongo packages
